+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes will be tracked in this file
 
 ##[Unreleased]
 ###Added
+- RadialPoints.py file for determining position of camera during calculations
 - addon_menu.py for setting up UI menu
+
 
 
 ##[0.1] - 2016-12-14